PSB-backend/PSB_PFM/requirements_backend.txt
asgiref==3.4.1
cycler==0.10.0
Django==3.2.7
django-rest-auth==0.9.5
djangorestframework==3.12.4
et-xmlfile==1.1.0
imageio==2.9.0
imutils==0.5.4
joblib==1.0.1
kiwisolver==1.3.2
matplotlib==3.3.4
networkx==2.6.3
numpy==1.21.2
opencv-python==4.5.1.48
openpyxl==3.0.8
pandas==1.3.3
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.2.0
tifffile==2021.8.30
